refactor: move Messages to constants package
